Revert "Organizando as camada .para tornar o código mais modular, escalável e fácil de manter"
