+++ src/modules/projects/dto/update-project.dto.ts
@@ -1,6 +1,7 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateProjectDTO } from '../dto/create-project.dto';
-
-export class UpdateUserDTO extends PartialType(CreateProjectDTO) {}
-
-export { CreateProjectDTO };
+export type ProjectDTO = {
+    id?: string;
+    title: string;
+    link: string;
+    tags: string;
+    description: string;
+}
