--- src/modules/projects/endpoints/get.projects.service.ts
@@ -1,11 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/database/PrismaService';
-
-@Injectable()
-export class GetProjectsService {
-  constructor(private prisma: PrismaService) {}
-
-  async findAll() {
-    return this.prisma.project.findMany();
-  }
-}
