+++ src/modules/projects/entities/project.entity.ts
@@ -1,9 +1,1 @@
-export class Project {
-   id?: string;
-   title: string;
-   tags: string;
-   description: string;
-   activeted: Boolean
-   createAt: Date;
-   updateAt: Date;
-  }
+export class Project {}
